+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 * Bugfix in volume-writing location, and force pretty formatting on written logs
 * Made the existing test actually query the node it created
 * Added another test to demonstrate an initial network setup
+* When asked about test suite data, send back a JSON of test suite metadata (rather than just a list of test names)